only allow delete user on dev version

+++ BrightID/src/components/HomeScreen.js
@@ -76,34 +76,35 @@ export class HomeScreen extends React.Component<Props> {
           title="options"
           iconName="dots-horizontal"
           onPress={() => {
-            // if (__DEV__) {
-            Alert.alert(
-              'WARNING',
-              'Would you like to delete user data and return to the onboarding screen?',
-              [
-                {
-                  text: 'Cancel',
-                  onPress: () => console.log('Cancel Pressed'),
-                  style: 'cancel',
-                },
-                {
-                  text: 'Sure',
-                  onPress: async () => {
-                    try {
-                      navigation.navigate('Onboarding');
-                      await AsyncStorage.flushGetRequests();
-                      await AsyncStorage.clear();
-                      store.dispatch(removeUserData());
-                    } catch (err) {
-                      console.log(err);
-                    }
+            if (__DEV__) {
+              Alert.alert(
+                'WARNING',
+                'Would you like to delete user data and return to the onboarding screen?',
+                [
+                  {
+                    text: 'Cancel',
+                    onPress: () => console.log('Cancel Pressed'),
+                    style: 'cancel',
                   },
-                },
-              ],
-              { cancelable: true },
-            );
-          }}
-          // }
+                  {
+                    text: 'Sure',
+                    onPress: async () => {
+                      try {
+                        navigation.navigate('Onboarding');
+                        await AsyncStorage.flushGetRequests();
+                        await AsyncStorage.clear();
+                        store.dispatch(removeUserData());
+                      } catch (err) {
+                        console.log(err);
+                      }
+                    },
+                  },
+                ],
+                { cancelable: true },
+              );
+            }
+          }
+          }
         />
       </HeaderButtons>
     ),
@@ -140,7 +141,7 @@ export class HomeScreen extends React.Component<Props> {
               source={{
                 uri: `file://${RNFS.DocumentDirectoryPath}/photos/${
                   photo.filename
-                }`,
+                  }`,
               }}
               style={styles.photo}
               resizeMode="cover"
